git: remove broken send-email script
git-send-email works only with user-installed perl (no longer supported)
or by installing modules on system perl (discouraged). Better to not
ship a script that we know will break: indeed, git.rb already removes
git-imap-send if openssl is not going to be present. Now do the same,
unconditionally, for git-send-email.

Fixes #33524.