llvm: build with Ninja instead of GNU Make
Fixes #35513. This commit adds Ninja as a build dependency, and
replaces the CMake Unix Makefiles generator/GNU Make build toolchain
with a CMake Ninja generator/Ninja toolchain.

An alternative fix proposed by @fxcoudert and tested by @ducky427
involves leaving the build toolchain as-is and setting the
`HOMEBREW_MAKE_JOBS` shell environment variable to `1`. This approach
does not introduce an additional build dependency, but is likely to
slow down a build from source significantly compared to the approach
implemented in this commit (all other things equal) because Ninja is
generally a faster build system, and the Ninja-based build system can
use all system cores versus using a single core in the Make-based
approach.