dmg2img 1.6.7 (new formula)
This adds a formula for dmg2img and vfdecrypt.  dmg2img is a CLI tool used to convert macOS DMG files to raw image files, and is used by some Linux scripts when dealing with DMG files.  vfdecrypt is a utility to decrypt encrypted DMG files if given the correct decryption key.