BUG:  CMake find_package(TBB) mechanisms were broken
Installing tbb in the homebrew prefered directory
layout caused cmake config based file management
to fail.

The bug was reported to upstream https://github.com/01org/tbb as
PR#119 to provide the flexibility needed to meet
the non-default (but intended modifiable) install location
of the TBBConfig.cmake file.

A temporary patch was added that implements the contents
of PR#119.

A set of tests was added to confirm that the desired
functionality also exists.