Remove x11 dependency
Pari/gp can be used on the command line without graphics. The x11
dependency results in error because homebrew does not automatically
install x11.

There is alternative graphics support through qt, but it needs to be
compiled with the "no-framework" flag (non default" and there is fltk,
which results in a compilation error.

I believe it is a good idea to just use it on the command line.