sqlite3-pcre (new formula)
This commit adds a new formula for sqlite3-pcre, an SQLite3 extension that uses
PCRE to provide an implementation of the SQL function REGEXP.

This formula attempts to address some of the concerns raised in
Homebrew/legacy-homebrew#26354. Rather than building the extension into sqlite
itself, it is provided as a separate formula that folks can install or not as
they please.

This package is a little unusual in that it installs a shared library (`.so`)
that can be dynamically loaded by SQLite in order to support the REGEXP
operator. This has to be done manually with the `.load` meta-command in SQLite,
and I note this in the caveats.

There are some issues with upstream that I'd like to acknowledge here, but I
hope that we can look past these since being able to help folks who need the
REGEXP operator is desirable and this seems to be the best way to do it.

   * Yes, this code is head-only. Upstream has never tagged a release. 🙄
   * Yes, it depends on a keg-only formula. The version of SQLite that ships
     with macOS doesn't support loading extensions, so it's necessary to install
     and use the Homebrew `sqlite` formula. This is mentioned in the caveats.
   * Yes, upstream is not very notable. I think in this case it's because the
     extension is basically done; there's no development happening because it's
     fairly simple and already just works.

The upstream Makefile assumes that `install` is GNU install, so I added a build
dependency on `coreutils` in order to provide that.