IXWebSocket 1.1.0
Add a formula for IXWebSocket, a WebSocket client and server, and HTTP
client. It comes with a command-line tool, ws, which is a simple
WebSocket toolkit. Tools include connect, chat, broadcast and echo
server, and send and receive tools to transfer files.  It also has a
curl like HTTP tool.