New formula for Paperspace CLI for cloud computing
This is a formula that uses node but builds a stand-alone binary that
doesn't require node to run.
Paperspace calls the github project "paperspace-node" and lists its
homepage as the Paperspace API homepage, but the CLI is the command-line
embodiment of the API.  Its specific URL:
https://paperspace.github.io/paperspace-node/#paperspace-cli
It is (as far as my communications with Paperspace support) the
reference CLI for managing and monitoring your Paperspace remote jobs.
Paperspace is a cloud compute service popular for running Machine
Learning compute jobs.
https://www.paperspace.com/