Add regex to strip time info returned from `psql`
The call to `psql` to get the database options expects a return value
that is just a flat string, something like: `en_US.UTF-8\n`. This fails
when a user has a custom `.psqlrc` file with this option on:

```pg
-- Show how long each query takes to execute
\timing
```

Instead, the output looks like this: `en_US.UTF-8\nTime: 1.471 ms`.
That, of course, is an invalid locale, which in turns causes the `initdb`
step to fail.

The fix is a simple regex that strips this info out.