ctags: fix build on recent macOS 10.14 - conflict with __unused__
The build failed on latest macOS 10.14.5 like so:

    In file included from main.c:62:
    /Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/include/dirent.h:80:2: error: use of undeclared identifier 'unused'
            __unused long   __padding; /* (__dd_rewind space left for bincompat) */
            ^
    /Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/include/sys/cdefs.h:161:40: note: expanded from macro '__unused'
    #define __unused        __attribute__((__unused__))
                                           ^
    ./general.h:60:37: note: expanded from macro '__unused__'
    # define __unused__  __attribute__((unused))
                                        ^
    1 error generated.
    make: *** [main.o] Error 1

As you can see, exuberant-ctags has defined __unused__ but the
macOS toolchain also uses __unused__ and it is confused.

The easiest fix is to pretend this is unsupported and disable/blank
`__unused__` (which should only affect compiler warnings, and the
same for the printf format string warnings which are also disabled).