tfenv: install share directory for gpg verification
tfenv has logic for verifying the Terraform releases that it downloads
with gpg or gpgv. When gpgv is used, tfenv tries to verify the
download's signature based on a key for Hashicorp that is bundled with
the tfenv release in the `share` directory - but the homebrew formula is
not installing that directory.