Update gpgme.rb
The PR submitted by the user for QT bindings didn't add them properly as dependencies and didn't exclude certain packages, such as QGPGME, which aren't even for this platform. This isn't even a dependency to begin with but optional. Until, it's properly added, then stop breaking GnuPG which is of more value than the QT bindings for a handful of formula.