Prefer homebrew-installed tcl-tk to system tcl-tk for git-gui
In order to do so, we replace the Wish binary that is used by git-gui with
a shell script that checks if the user has install tcl-tk with homebrew
and if so uses that to launch git-gui instead of the system tcl-tk.

Without this change, all dialogs that are shown by git-gui are blank
windows and can only be closed by memorizing the location of the buttons.

This change does not add a dependency on tcl-tk for git, it only ensures
that git-gui works better if tcl-tk is installed. git-gui will still launch
if tcl-tk is not installed but it will use the broken system-provided wish
binary. Once tcl-tk is installed with homebrew, we pick the homebrew wish
binary instead and git-gui will work as expected.

This fixes #36390 and #39987