Move bin/python to pkgshare/python
The python directory and the two python scripts inside that directory are moved from bin/ to pkgshare/  because the python scripts require python module "tensorflow" and that requires homebrew formula "hdf5" to install correctly. Being as the ffmpeg make file doesn't do this itself, it seems that these files are at most experimental or at least examples.