pass: improve speed when calling gnu-getopt
The `pass` tool requires `gnu-getopt` which on macOS is done by (amongst others) trying:

    brew --prefix gnu-getopt

This has two downsides:

1. It is slow, see e.g. Homebrew/brew#3097 (0.5-0.8s).
2. It can break, see e.g. https://lists.zx2c4.com/pipermail/password-store/2019-September/003774.html

For this reason `pass` should not use `brew` and this patch changes the relevant line to use an absolute path.

This can’t be fixed upstream, as the location of `gnu-getopt` depends on which package manager installed `pass`.