Add tac Formula (new formula)
This PR adds a new formula `tac`, since macOS dose not include this for
some reason, it will be really nice to have this here!

This is a quick little tac program I wrote up, should work as a
substitute for your Linux tac command.

_I'm very new to brew, (and I'm liking it!), never used ruby before,
hopfully I did everything correctly..._