Allow pony related bottles to run on CPUs older than the build machine
By default, all Pony project makefiles optimize for the CPU of the machine
being built on unless the `arch` option is provided. Today, we had a user
with an older machine report that the Homebrew bottles didn't work for them.
They were encountering an illegal instruction exception.

This patch fixes all 3 Pony related formulas to stick to a basic x86-64 instruction
set.