Removing awscli.
Removing awscli as the new version has moved this from being a python
package to a macOS pkg distributed by AWS with bundled python and app
the dependencies it needs.

A new cask has been created in homebrew/homebrew-cask#76967 and this
should only be merged once that PR has been accepted.