Upgrading awscli to v2.0.0
Upgrading the aws cli to version 2.0.0 from source.

There are already two PRs for this, my previous #50056 which was to
remove this and replace with a Cask which was rejected and #50029 which
has a broken build.

This changes the pip install like for the requirements to allow
installing the package of botocore as the upstream has changed the
dependencies in requirements.txt.

Tested the build and it appears to work.