zpython: delete
Depends on EOL Python 2

install: 13 (30 days), 56 (90 days), 162 (365 days)
install-on-request: 13 (30 days), 56 (90 days), 158 (365 days)

Looks abandoned since 5 years: https://bitbucket.org/ZyX_I/zsh/commits/
Does not work with Python 3
Ships an outdated zsh version, and uses a head build provided by a gist
file ... this formula would probably not be accepted as-is in
homebrew-core nowadays.

As this is a low-usage formula and we do not want to maintain it
further, let's remove it. Users can still use `brew extract` and
move it to a personal tap if needed.