qu-tree 0.1 (new formula)
A new formula for QuTree, a tensor tree linear algebra focused on quantum dynamics and machine learning.
Passes brew test, brew audit --strict --online, and brew audit --new-formula with the exception of not being notable enough.
As a counterpoint to notability, this package is used internally at multiple academic institutions and is our first step in refactoring and migrating these tools to open-source.
Several publications using the core functionality of this package
already exist and more downstream projects are planned for this year,
so having an installable package would greatly clarify our dependencies.