ruby: create symlink to Gem binaries in unversioned path
Currently, the gem binaries are install at path like `/usr/local/lib/ruby/gems/2.7.0/bin`,
which includes the ruby version number. This has the disadvantage that users are required to
manually update their PATH environment whenever there is a major ruby release.

To fix this, we create a symlink in prefix to point to the `rubygems_bindir`.
As such, users can instead add `/usr/local/opt/ruby/gembin` to the PATH.

This change is also compatible with existing setup.