Update Tarantool main release
Set stable release to 2.5.2.0 version.

Not to depend on a backend storage we use, changed target path to:

  download.tarantool.org

Met issue with building on OSX 13:

  src/lua/init.c:478:2: error: use of undeclared identifier 'rl_catch_signals'
        rl_catch_signals = 0;
        ^
  src/lua/init.c:479:2: error: use of undeclared identifier 'rl_catch_sigwinch'
        rl_catch_sigwinch = 0;
        ^

Found that dependencies headers are included into the default search
paths and SDK's headers are not included into the default paths. To
fix the issue ncurses header and library were set explicitly in the
Tarantool formula:

    args << "-DCURSES_NEED_NCURSES=TRUE"
    args << "-DCURSES_NCURSES_INCLUDE_PATH=/usr/include"
    args << "-DCURSES_NCURSES_LIBRARY=/usr/lib/libncurses.dylib"

Also found that the similar problem appears for iconv and the
following was also added to formula:

    args << "-DICONV_INCLUDE_DIR=/usr/include"

Disabled use of bundled libcurl:

    args << "-DENABLE_BUNDLED_LIBCURL=OFF"

after it removed autotools from dependencies and set 'curl' instead
using uses_from_macos routine, also set for 'ncurses'.

Closes tarantool/tarantool#5322

Co-authored-by: Alexander Turenko <alexander.turenko@tarantool.org>