lc0 0.25.1 (new formula)
Add a new formula for Leela Chess Zero (`lc0`). Leela Chess Zero is
one of the strongest chess engines and is interesting because it is
neural-network-based (most strong chess engines aren't). This formula
compiles the `lc0` binary and then downloads a network file to use.

This formula uses a medium-size network as recommended in the project
documentation. Read more here:

https://github.com/LeelaChessZero/lc0/wiki/Best-Nets-for-Lc0

This formula is based on that of `leela-zero`, a similar engine for the
game Go.