r: fix build by installing gss using install.packages
The gss package is at times a moving target and as a result the URL
for the resource is not stable. This fix changes the test to let R
handle download and installation of gss so that the release of new
versions of gss won't cause the test to fail due to link breakage.