sloccount: remove redundant dependency on md5sha1sum
sloccount can use macOS's system `/sbin/md5` command, and in fact will do so if
neither md5sha1sum nor coreutils are installed. The only problem is that it
checks for an `md5sum` command first, and prints an annoying warning if it can't
find one.

This patch does two things:

   * removes unneeded dependency on md5sha1sum, and
   * patches `break_filelist` to suppress the warning.

md5sha1sum conflicts with coreutils because both install an `md5sum` command
(among other things). sloccount is the only package in homebrew/core that
depends on md5sha1sum. This patch eliminates that.