Macaulay2 1.16 (new formula)
Adds a brew formula for Macaulay2, a software system for algebraic geometry research.

The formula uses the CMake build for Macaulay2, which begins with building a number of library and program dependencies not available via brew. After building Macaulay2, the examples are executed and the documentation is generated. In total, the build can up to two hours.