file-formula: bump revision to pick up the latest libmagic
`file-formula` was built once when `libmagic` was still 5.38.
Need a revision bump to pick up the latest libmagic.

Otherwise, we will have:

```sh
$ file

file: Compiled magic version [538] does not match with shared library magic version [539]
```

In the future, it would be nice to bump both `libmagic` and
`file-formula` in a same pull request.