libsharp 1.0.0 (new formaula)
This library is distributed as part of healpix, but needs to be build seperatly as of healpix 3.60.
The current formula for healpix builds but does not install libshap leading to unresolved symbols.
Since healpix needs to find libsharp at build time the perfered solution is to install it as a seperate
formula. Once this formula is accepted I will modify healpix's formula to build against this version of libsharp.