gcc: Fix `version_suffix`
The original implementation was fundamentally broken when called from other formulae in HEAD builds.

Per the discussion in https://discourse.brew.sh/t/calling-one-formulas-methods-in-another/8382, I've renamed the original method to `buildver_suffix` for internal use only, and defined a new `version_suffix` that does the right thing.

The `version_suffix` comment is an implied acknowledgement that this method is for "public use".