sponge 0.63 (new formula)
sponge is part of moreutils, which is already present in Homebrew.
However, sponge by itself is quite simple and has no dependencies
outside of the C standard library, and is a very useful utility on its
own. moreutils additionally conflicts with a few other common packages
(parallel being the most obvious) and requires a specific tap to exclude
those from install.