cdktf remove depenecy on terraform
I use tfenv to manage multiple installations/versions of terraform.
And I'd like to use homebrew to install cdktf, but it has a hard
depenecy on terraform. This PR removes that depenency.

I think it would be better to have some sort of optional or
recommended depency on terraform or tfenv, but I don't see a way
to do that with homebrew - please correct me if I'm wrong.