Inko 0.8.1
This release fixes the build errors noted in
https://github.com/Homebrew/homebrew-core/pull/62035.

Version 0.8.0 (which doesn't build on macOS) introduced a new build
system, and added support for using a system installation of libffi.
These changes mean that automake, autoconf, and libtool are not
required; in exchange for making libffi a runtime dependency. This also
makes it possible to update libffi in the event of security updates,
without having to recompile Inko.