Added xianyi/OpenBLAS#2729 patch and fixed TARGET
The currect version of openBLAS causes numpy to crash
on HASWELL or later architectures when working with
large matrices.  The first change is to patch openBLAS
to fix this error.  According to the openBLAS developers
this will be in 0.3.11 and so the patch can be removed
then.

I also adjusted the TARGET according to Jon Chang's
suggestion so that it will build on the oldest
possible architecture on both x86 and ARM. With
DYNAMIC_ARCH=1 openBLAS should now work on all
architectures.