Update mono.rb
Unsure if this is all that's required for python3 dependency. It may require an include and virtualenv for the compiling or execution later. Not sure if mono links python3 for calling and not sure about the python@3.9 being keg only is going to require some sort of linking for the *.py.