isl@0.18 (new formula)
This is needed for gcc@5 on linux.
I tried to use isl 0.18 built as a resource within the gcc@5 formula
but the gcc binaries are still linked to the system isl, whatever flags
or options I tried to set.
Using isl installed as a formula produces gcc@5 binaries that are linked
to that isl, so this is the only solution I found to make this work.