emacs.rb: Add support for Apple silicon
Apple requires all binaries to be signed in order for them to be valid
on M-series Macs. emacs modifies some of its binaries after they are
created which invalidates the signature. This change to the emacs
formula hacks on a codesign -s - after the modification to ensure the
binary can be run. Will work with upstream to see about a better fix
but this get things running.

Signed-off-by: Nathan Hjelm <hjelmn@google.com>