netcdf-fortran 4.5.2 (new formula)
netcdf-fortran was part of the homebrew formula netcdf. I have removed it from there and now commit a new formula for netcdf-fortran only, which depends on netcdf. This is to avoid conflicts with other netcdf-fortran installations, e.g. with Intels ifort, that are using homebrew's netcdf-c installation. This is also the reason why it is keg_only.