emscripten: cleanup formula
The emscripten depends on a non-stable build of llvm, and is therefore
packaged with with the required llvm build.

This commit does the following:
- removes unnecessary build targets in the llvm project
- adds a test that the llvm dependencies were in fact built
- declares a dependency on libffi [*]
- makes llvm build commands consistent with llvm formula

Until https://github.com/Homebrew/homebrew-core/pull/63183, the
emscripten installation used nearly 2GB of storage space. Some
improvements were made in that PR, and this commit is meant to follow
through on that.

[*] The arguments to cmake already previously referred to the libffi
formula. It wasn't declared as a dependency, so the llvm install linked
with the built-in libffi.