lua@5.3: fix pc file
The generated pc file points to `lua[@5.4]`, not `lua@5.3`. This commit
fixes that. After this change:

    ❯ export PKG_CONFIG_PATH="/usr/local/opt/lua@5.3/lib/pkgconfig"
    ❯ pkg-config --libs --cflags lua5.3
    -I/usr/local/opt/lua@5.3/include/lua -L/usr/local/opt/lua@5.3/lib -llua.5.3 -lm

One caveat is that adding `lua@5.3` to `PKG_CONFIG_PATH` now means that
using `pkg-config` to generate build flags for lua without specifying a
version will point to `lua@5.3`. I'm not hugely bothered by this, since
`lua@5.3` is keg-only.