make zeromq depend on libsodium
back when Formulae had options, zeromq dependency on libsodium was recommended,
then disabled due to a build issue in ntopng.

I believe this has been fixed in more recent versions of ntopng,
so we can restore the zeromq 'recommended' behavior as the only behavior for zeromq.

ntopng dynamically links libsodium, unlike it does with zeromq