install clingo python module to loadable path
The `-DPYCLINGO_USER_INSTALL=OFF` setting needs to be passed to get the
library to be placed in the global site-packages.  Without it, it gets
installed into a strange home-directory path that is not part of any
normal python search path.

Also, the variable used to set the executable path for FindPython is
Python_EXECUTABLE, rather than PYTHON_EXECUTABLE.