llvm: restrict openmp ARM patch to stable version
The relevant upstream patch was subsumed in:
https://reviews.llvm.org/D88252

This patch was landed 2020-11-24:
https://github.com/llvm/llvm-project/commit/7b5254223acbf2ef9cd278070c5a84ab278d7e5f

It had a bug that was fixed 2020-11-25:
https://github.com/llvm/llvm-project/commit/9e3e332d273b80b5167ac35f8dcfa7178e45c5e9

After these commits, the patch won't apply and isn't necessary.
Restricting it to the stable source fixes a build error with `--HEAD`.
I still can't get the `--HEAD` build to complete successfully on either
x86 or ARM because of an unrelated error during compilation, but this
gets much further than before.

Moved up above the livecheck to satisfy the `audit` results.