trader: add license
From COPYING:

============================================
Star Traders: A Game of Interstellar Trading
============================================

**Copyright © 1990–2020, John Zaitseff.**

The Star Traders game is free software that is distributed under the
terms of the GNU General Public License.  The actual copyright on this
program (“the Program”, in the language of the License) belongs to John
Zaitseff, although the text of the License itself belongs to the Free
Software Foundation.

Even though the GNU General Public License does *not* require you to send
your modifications back to the author, it is considered “good form” to do
so, as this allows your modifications to be incorporated into future
versions of the program, allowing others to benefit from them.

==========================
GNU GENERAL PUBLIC LICENSE
==========================
--------------------------
 Version 3,  29 June 2007
--------------------------

| Copyright © 2007 Free Software Foundation, Inc.
| https://fsf.org/

    *Eeryone is permitted to copy and distribute verbatim copies of this
    license document, but changing it is not allowed.*

From lib/c++defs.h:

/* C++ compatible function declaration macros.
   Copyright (C) 2010-2020 Free Software Foundation, Inc.

   This program is free software: you can redistribute it and/or modify it
   under the terms of the GNU General Public License as published
   by the Free Software Foundation; either version 3 of the License, or
   (at your option) any later version.

   This program is distributed in the hope that it will be useful,
   but WITHOUT ANY WARRANTY; without even the implied warranty of
   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
   General Public License for more details.

   You should have received a copy of the GNU General Public License
   along with this program.  If not, see <https://www.gnu.org/licenses/>.  */