graphviz 2.46.0
The latest release isn't available at the original URL. However, the
graphviz website download page [1] links to GitLab.

We use a git checkout for the source because the autogen.sh script looks
for a git repository to recover version information.

The livecheck block has been removed. It is no longer needed given the
GitLab URL.

[1] https://www.graphviz.org/download/source/