git-interactive-rebase-tool 2.0.0
This update was non-simple as the existing tests were broken by changes
in the project. The tests failed because the existing tests depend on
reading input from stdin, but this project no longer supports reading
from stdin in all cases. To fix this, providing a noop rebase file will
allow the application to run and successfully exit, without the need to
virtually interact with the program.

Also, the project no longer depends on ncurses, so that dependency has
been removed.