idris2: add runtime dependency on coreutils
The `idris2` that is generated is essentially a very tiny script
that uses `realpath` at startup to figure out what `LD_LIBRARY_PATH`
to set before running the executable that chez produced.

So we need `coreutils` for the build but also to actually run the
executable.

I experienced this recently while setting up a macOS CI and I have
just noticed that it had been reported to us previously:
https://github.com/idris-lang/Idris2/issues/874