inspircd: improve the flags passed to configure.
--disable-auto-extras disables automatically detecting the extra
modules to use. This is necessary when building a package to avoid
hidden dependencies on installed libraries.

--distribution-label helps upstream diagnose package issues.

Note: the regex_posix and sslrehashsignal modules were previously
built automatically as --disable-auto-extras was not passed. This
commit preserves the way it was built previously for compatibility
purposes. Neither of these two modules have any non-POSIX deps.