Update asdf.rb
Switch to `opt_libexec` which translates to `opt_prefix/"libexec`, this means that there will be no migration steps for shims. 
However till adding a `$ asdf reshim` `post_install` hook incase there is an update that needs to be done on the shims when `asdf` itself updates to a new version or formula revision. There are changes pending on  [`asdf-vm`](https://github.com/asdf-vm/asdf/pull/893) that would allow regenerating the shims in case they need to be.