Set deprecation date for ruby@2.6
The release of Ruby 2.6.7† started the last phase of Ruby 2.6 life - the
security maintenance phase. The security maintenance phase is currently
scheduled for a year since the 2.6.7 release, therefore end of life of
Ruby 2.6 is scheduled for 2022-04-05. It usually doesn’t happen
precisely on the scheduled day, but we want to keep this date around.

† https://www.ruby-lang.org/en/news/2021/04/05/ruby-2-6-7-released/