llvm: Collapse runtimes and projects being configured.
Using the separate lists here isn't the documented way of building LLVM,
nor what is used by the upstream packagers. It results in different
directory layouts on different Linux distributions that don't work as
well with Clang and other instructions (and don't match the layout on
Mac). It also caused some other problems when porting to linuxbrew that
resulted in disabling things there.

Switching to the simpler all-projects configuration, which is the
documented and recommended way of building LLVM, these problems go away
and the predictable and usable. This causes the tests to pass, etc.

Everything also continues to work on Mac, and this really seems to be
the canonical and preferred way to build LLVM. I have built LLVM from
source with this on my Mac and ensured the tests pass and the resulting
toolchain works.

I also have tested a version of this in Linuxbrew on both a Debian
testing and Ubuntu 16.04 (the distro baseline where historical problems
occurred) and it works well and passes all tests.

If anyone hits issues with LLVM builds in Homebrew, I'm happy to help
debug and help maintain the LLVM build here. I'm a very long time LLVM
developer and user with lots of experience with the build system.