gawk: skip make check if no cmp command is installed
We have a few users on Linux who compile from source,
and did not have cmp installed on their system.
As gawk is quite low in the dependency tree, it's important
to be able to install it. The tests can be skipped locally,
they are still run on macOS and Linux CI, where cmp is installed