fltk: add head url and install steps
The newer releases of fltk have different build & install steps, so make
the install method follow whichever is applicable.

Once 1.4.0 is mainline stable, we can remove the old steps, and make
cmake a general build dependency, not just for the head.

Also added the license while I was editing things -- FLTK uses the
LGPLv2 with an exception recognized by SPDX.