haskell-language-server: fix test, prepare ARM build
The current formula encodes the full version of GHC in the name of the
binary. This leads to a test failure in GHC version bump PRs (cf. #78821).
Let's fix that by using only the major and minor versions.

While we're here, let's get this formula ready for bottling for ARM.