klee: switch to LLVM 12
Let's apply the relevant patches from klee/klee#1389 to enable support
for LLVM 12.

While we're here, let's enable support for STP now that it's available
in Homebrew/core. STP is the default solver used by klee. I've also
reorganised the build flags a little to make it easier to read.