pillow 8.2.0 (new formula)
This is a Python library, but it has no Python dependencies, and could
be useful for other formulae.

This formula fails the new formula audit because of the `tcl-tk`
dependency. However, some formulae that use Pillow also depend on
`tcl-tk`, and we don't want to have a mixed-version dependency when we
switch those formulae to depend on this one.

Partially addresses #79562. See also #73875.