workflows: enable Linux bottling for a limited amount of packages
This enables bottlings on pull requests for the formulae defined in
linux-pr-tests-yml only.

The difference between both core repos is now down to 115 formulae
that still need a backported. We can already start slowly bottling
stuff, and work on backporting in parallel. I estimated that
we could bottle around 1500 to 2000 formulae right now, but I would
like to focus on the main 100 formulae to start with.

The strategy I would like to use is the following:

- Merge this PR
- Use the linux-dispatch-build-bottle workflow to build around 100 low-level bottles (zlib, bzip2, ...): basically the formulae that are needed to build the linux only formulae.
- Add these 100 formulae to the linux-pr-tests workflow: the current linux-only bottles depend on these
low-level formuale and should be bottled at that point, making the list self-consistent.
This is done to comply with the request to start bottling in a very controlled way: https://github.com/Homebrew/homebrew-core/pull/72732
- At this point we do not need the self-hosted runner workflow yet, as builds should last less than 6 hours

Once all the linux-only formulae are bottled (with their dependencies),
we will proceed in a second phase by adding the self-hosted runner workflow, and start bottling even more.

In parallel of this, we will continue to finish the backporting of the remaining 115 formulae.
At the current pace this could still take between 2 to 4 weeks.