llvm-bootstrap 12.0.1 (new formula)
This splits the bootstrap stage from the `llvm` formula into a separate
formula. This will allow us to save CI time by "caching" the bootstrap
stages and skipping them when they're not needed.

For example, this should shave a few hours off of CI runs from PRs that
bump LLVM's `revision` because of a dependency version bump (cf. #80227).