workflows/tests: enable use of a `CI-force-dependents-tests` label
Currently, CI skips dependents if the formula tests fail for any reason.
This can be undesirable in PRs which require a large number of revision
bumps (e.g. `icu4c`, `boost`, or `libffi`), or when failures are due to
transient networks issues or other issues that can be more easily
addressed in separate PRs (e.g. some `livecheck` failures).

This should allow us to make use of a label that will force CI to
continue attempting to test dependents even when the tests of the
modified formula fails.