llvm: various improvements
1. Add patch to fix parallel builds on ARM.
2. Clean up some duplication
3. Install the `lit` command line tool
4. Install Vim editor integration

The `lit` tool adds very little to the bottle size, and it seems some
users might find it useful. Let's install it directly using Python to
avoid having to set `LLVM_INCLUDE_TESTS` to `ON`, as that has many other
side-effects that we don't need or want. This is also how Arch installs
`lit`. [1]

Closes #82479.

[1] https://github.com/archlinux/svntogit-packages/blob/5e7a8f09f53d0fb19c0680bb5b4691b1c6a78ad1/trunk/PKGBUILD#L79-L82